+++ src/types.ts
@@ -101,6 +101,17 @@ export function fromProgramOrderType(orderType: any): OrderType {
     return OrderType.POSTONLYFRONT;
 }
 
+export enum SelfTradeBehavior {
+  CANCELPROVIDE,
+  ABORTTRANSACTION,
+}
+
+export function toProgramSelfTradeBehavior(behavior: SelfTradeBehavior) {
+  if (behavior == SelfTradeBehavior.CANCELPROVIDE) return { cancelProvide: {} };
+  if (behavior == SelfTradeBehavior.ABORTTRANSACTION)
+    return { abortTransaction: {} };
+}
+
 export enum Side {
   BID,
   ASK,
@@ -598,6 +609,7 @@ export interface OrderOptions {
   clientOrderId?: number;
   tag?: string;
   blockhash?: { blockhash: string; lastValidBlockHeight: number };
+  selfTradeBehavior?: SelfTradeBehavior;
 }
 
 export interface TriggerOrderOptions {